contributor-stats/pyproject.toml
[tool.poetry]
name = "contributor-stats"
version = "0.1.0"
description = "Generates statistics for contributors from git history and GitHub activity."
authors = ["Erik Bjäreholt <erik@bjareho.lt>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "*"
PyGithub = "*"
numpy = "*"
tqdm = "*"
click = "^8.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
mypy = "*"
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
